Add files via upload
Instantiates effects for stun, shield and battery hits... leave behind gameobject of the particle effect unfortunately, wouldn't let me destroy it easily without changing all the code that we have and i don't have time. maybe we can find a way around it when i get to uni to talk with Ryan. unless Ryan can figure it out. (the game objects are attached to the laser prefab) and dont destory when the laser destroy's itself for some reason